+++ fastai/basic_data.py
@@ -3,7 +3,7 @@ from .torch_core import *
 from torch.utils.data.dataloader import default_collate
 
 DatasetType = Enum('DatasetType', 'Train Valid Test Single Fix')
-__all__ = ['DataBunch', 'DeviceDataLoader', 'DatasetType']
+__all__ = ['DataBunch', 'DeviceDataLoader', 'DatasetType', 'load_data']
 
 old_dl_init = torch.utils.data.DataLoader.__init__
 
@@ -138,7 +138,22 @@ class DataBunch():
 
     def add_tfm(self,tfm:Callable)->None:
         for dl in self.dls: dl.add_tfm(tfm)
+            
+    def remove_tfm(self,tfm:Callable)->None:
+        for dl in self.dls: dl.remove_tfm(tfm)
+            
+    def save(self, fname='data_save.pkl'):
+        if not getattr(self, 'label_list', False):
+            warn("Serializing the `DataBunch` only works when you created it using the data block API.")
+            return
+        torch.save(self.label_list, self.path/fname)
 
+    def add_test(self, items:Iterator, label:Any=None)->None:
+        self.label_list.add_test(items, label=label)
+        vdl = self.valid_dl
+        dl = DataLoader(self.label_list.test, vdl.batch_size, shuffle=False, drop_last=False, num_workers=vdl.num_workers)
+        self.test_dl = DeviceDataLoader(dl, vdl.device, vdl.tfms, vdl.collate_fn)
+        
     def one_batch(self, ds_type:DatasetType=DatasetType.Train, detach:bool=True, denorm:bool=True, cpu:bool=True)->Collection[Tensor]:
         "Get one batch from the data loader of `ds_type`. Optionally `detach` and `denorm`."
         dl = self.dl(ds_type)
@@ -242,3 +257,10 @@ class DataBunch():
             except: pass
             warn(message)
             print(final_message)
+
+def load_data(path:PathOrStr, fname:str='data_save.pkl', bs:int=64, val_bs:int=None, num_workers:int=defaults.cpus, 
+                  dl_tfms:Optional[Collection[Callable]]=None, device:torch.device=None, collate_fn:Callable=data_collate, 
+                  no_check:bool=False, **kwargs)->DataBunch:
+    ll = torch.load(Path(path)/fname, map_location='cpu') if defaults.device == torch.device('cpu') else torch.load(Path(path)/fname)
+    return ll.databunch(path=path, bs=bs, val_bs=val_bs, num_workers=num_workers, dl_tfms=dl_tfms, device=device,
+                        collate_fn=collate_fn, no_check=no_check, **kwargs)