+++ fastai/data_block.py
@@ -496,6 +496,7 @@ class LabelLists(ItemLists):
         if getattr(self, 'normalize', False):#In case a normalization was serialized
             norm = self.normalize
             data.normalize((norm['mean'], norm['std']), do_x=norm['do_x'], do_y=norm['do_y'])
+        data.label_list = self
         return data
 
     def add_test(self, items:Iterator, label:Any=None):