+++ tests/test_text_data.py
@@ -123,13 +123,13 @@ def test_load_and_save_test():
     df = text_df(['neg','pos'])
     data = TextClasDataBunch.from_df(path, train_df=df, valid_df=df, test_df=df, label_cols=0, text_cols="text", bs=10)
     data.save()
-    data1 = TextClasDataBunch.load(path, bs=10)
+    data1 = load_data(path, bs=10)
     assert np.all(data.classes == data1.classes)
     assert np.all(data.train_ds.y.items == data1.train_ds.y.items)
     str1 = np.array([str(o) for o in data.train_ds.y])
     str2 = np.array([str(o) for o in data1.train_ds.y])
     assert np.all(str1 == str2)
-    shutil.rmtree(path/'tmp')
+    os.remove(path/'data_save.pkl')
 
 def test_sortish_sampler():
     ds = [1,2,3,4,5,6,7,8,9,10]