+++ CHANGES.md
@@ -15,7 +15,19 @@ of that change.
 
 
 
-## 1.0.44.dev0 (Work In Progress)
+
+
+## 1.0.45.dev0 (Work In Progress)
+
+### New:
+
+### Changed:
+
+### Fixed:
+
+
+
+## 1.0.44 (2019-02-13)
 
 ### New: